Split negotiation module into smaller files
